FEM: unit tests, split common module into a new module for result tests
